solve pr comments
