+++ engine/k8s-manager/config.yml
@@ -15,22 +15,3 @@ fluentbit:
   image: fluent/fluent-bit
   tag: 1.3
 configPath: /kai
-# nats_streaming:
-#   url: "kai-nats:4222"
-#   host: "kai-nats"
-#   port: "4222"
-#   storage:
-#     className: "standard"
-#     size: 1G
-# mongo:
-#   persistentVolume:
-#     storageClass: "standard"
-#     size: 5G
-# chronograf:
-#   persistentVolume:
-#     storageClass: "standard"
-#     size: 1G
-# influxdb:
-#   persistentVolume:
-#     storageClass: "standard"
-#     size: 5G
